requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
httpx==0.27.0
loguru==0.7.2

SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
pgvector==0.2.5
asyncpg==0.29.0

pymupdf==1.24.4
pdfplumber==0.11.0
pypdf==4.2.0
unstructured==0.14.10
pytesseract==0.3.10
Pillow==10.3.0

sentence-transformers==2.7.0
torch==2.3.1
numpy==1.26.4
scikit-learn==1.4.2

openai==1.35.7
anthropic==0.34.2
litellm==1.35.5

pytest==8.2.0
pytest-asyncio==0.23.6
coverage==7.5.1
ruff==0.5.5
black==24.4.2
mypy==1.10.0
pre-commit==3.7.1
